reformat: run ruff formatter and checker

+++ boostdm/src/discovery_index/muts.py
@@ -1,4 +1,3 @@
-
 from os import path
 
 import click
@@ -9,24 +8,24 @@ def build_table(files):
     """Build a table with mutations from multiple cohorts."""
     mutations_table = []
     for file in files:
-        cohort_name = path.basename(file).split('.')[0]
-        df = pd.read_csv(file, sep='\t')
-        df['COHORT'] = cohort_name
+        cohort_name = path.basename(file).split(".")[0]
+        df = pd.read_csv(file, sep="\t")
+        df["COHORT"] = cohort_name
         mutations_table.append(df)
     df = pd.concat(mutations_table, axis=0)
-    df['chr'] = df['chr'].astype(str)
-    df['pos'] = df['pos'].astype(int)
+    df["chr"] = df["chr"].astype(str)
+    df["pos"] = df["pos"].astype(int)
     return df
 
 
 @click.command()
-@click.argument('files', nargs=-1)
-@click.option('--output', type=click.Path())
+@click.argument("files", nargs=-1)
+@click.option("--output", type=click.Path())
 def cli(files, output):
     """Prepare mutation datasets for discovery step."""
     df = build_table(files)
-    df.to_csv(output, sep='\t', index=False, compression='gzip')
+    df.to_csv(output, sep="\t", index=False, compression="gzip")
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     cli()
